in tides.rs{implement the tidal_torqu_kaula_colplanar and the radial_component_tial_force_kaula_cpoplanar}= Compilation Nowarning -> execution fail error: thread 'main' panicked at 'called  on an  value: Unknown integrator'

+++ src/particles/particle.rs
@@ -77,7 +77,7 @@ impl Particle {
         let love_number_excitation_frequency: [[f64;32];32] = [[0.;32];32];
         let real_part_love_number: [[f64;32];32] = [[0.;32];32];
         let imaginary_part_love_number: [[f64;32];32] = [[0.;32];32];
-        let num_datapoints: f64 = 0. ;
+        let num_datapoints: i32 = 0;
         // ///////////////////////////////////////////////////////////////////////////////////////////////////
         let k_factor = 0.;
         let rotation_saturation = 0.;
@@ -132,7 +132,7 @@ impl Particle {
         let love_number_excitation_frequency: [[f64;32];32] = [[0.;32];32];
         let real_part_love_number: [[f64;32];32] = [[0.;32];32];
         let imaginary_part_love_number: [[f64;32];32] = [[0.;32];32];
-        let num_datapoints: f64 = 0.;
+        let num_datapoints: i32 = 0;
         // /////////////////////////////////////////////////////////////////////////////////////////
         let k_factor = 0.;
         let rotation_saturation = 0.;
